Update error tests
